+++ starlight/src/awi_structs.rs
@@ -1,10 +1,14 @@
+mod bridge;
 pub mod epoch;
 mod eval_awi;
+mod inout;
 mod lazy_awi;
 mod temporal;
 
+pub use bridge::Drive;
 pub use epoch::{Assertions, Epoch, SuspendedEpoch};
 pub use eval_awi::EvalAwi;
+pub use inout::{In, Out};
 pub use lazy_awi::LazyAwi;
 pub use temporal::{delay, Loop, Net};