+++ starlight/src/utils/ortho.rs
@@ -1,5 +1,8 @@
 use core::fmt;
-use std::ops::{Index, IndexMut, Neg};
+use std::{
+    array,
+    ops::{Index, IndexMut, Neg},
+};
 
 /// Represents an orthogonal direction on a grid
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
@@ -98,6 +101,10 @@ impl fmt::Display for Ortho {
 pub struct OrthoArray<T>(pub [T; 4]);
 
 impl<T> OrthoArray<T> {
+    pub fn from_fn<F: FnMut(Ortho) -> T>(mut cb: F) -> Self {
+        Self(array::from_fn(|i| cb(Ortho::try_from_usize(i).unwrap())))
+    }
+
     pub fn get(&self, ortho: Ortho) -> &T {
         &self.0[ortho.to_usize()]
     }
@@ -118,6 +125,33 @@ impl<T> Index<Ortho> for OrthoArray<T> {
 impl<T> IndexMut<Ortho> for OrthoArray<T> {
     fn index_mut(&mut self, i: Ortho) -> &mut T {
         self.get_mut(i)
+    }
+}
+
+impl<'a, T> IntoIterator for &'a OrthoArray<T> {
+    type IntoIter = std::slice::Iter<'a, T>;
+    type Item = &'a T;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.iter()
+    }
+}
+
+impl<'a, T> IntoIterator for &'a mut OrthoArray<T> {
+    type IntoIter = std::slice::IterMut<'a, T>;
+    type Item = &'a mut T;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.iter_mut()
+    }
+}
+
+impl<T> IntoIterator for OrthoArray<T> {
+    type IntoIter = <[T; 4] as IntoIterator>::IntoIter;
+    type Item = T;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.into_iter()
     }
 }