+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Additions
 - `Loop`s and `Net`s can now have an arbitrary initial value
 - Added `Grid`, `Ortho`, and `OrthoArray`
+- Added `In` and `Out` const width `RNode` structs
 
 ## [0.3.0] - 2024-01-22
 ### Fixes